==> 04_latex_results_generation.ipynb <==
import pandas as pd
import seaborn as sns
import numpy as np
import matplotlib.pyplot as plt

datasets_info = pd.read_csv('r_scripts/datasets/datasets_summary.csv')
datasets_info.head()

print(datasets_info[['Name','Nodes','Edges','Diameter','Modularity']].to_latex())

print(datasets_info[['Name','Density','Clustering.Coefficient','Betweeness','Closeness','Connected.Components','Communities']].to_latex())

results = pd.read_csv('experiment_results.csv')
results = results[results['method'] != 'Louvain-igraph']

results.head()

results['dataset'].unique()

# This for cycle is for saving images
for metric in ['P*/P','NMI']:
    for dataset in ['Dancer 01', 'Dancer 03', 'Cora']:
        for modification in results['modification'].unique():
            data = results[(results['metric'] == metric) &
                           (results['modification'] == modification) &
                           (results['dataset'] == dataset)]

            columns = ['percentage']
            columns.extend(data['method'].unique())
            
            data = data[['percentage','method','value']]
            data = data.pivot(index=['percentage'], columns=['method'])
            data.columns = ['_'.join(col).replace('value_','') for col in data.columns.values]
            data = data.reset_index()
            print('---------------------------------------------------------------------------------')
            print(metric,modification,dataset)
            print('\\begin{table}[H]')
            print(data.to_latex(index=False, float_format="{:0.3f}".format))
            print('\caption{Change of $P^{*}/P$ with random removal for '+dataset+'.}')
            print('\\end{table}')
            print('---------------------------------------------------------------------------------')


data.columns
