==> 09_decision_tree.ipynb <==
import os
import sys
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn import tree
from sklearn.svm import OneClassSVM
from sklearn.metrics import f1_score

dataset_source='D:/Pablo/clases/UJM/2. Semester, 2021/Mining Uncertain Social Networks/Repository/Experiments/datasets/dancer_01/'

results_rows = []
for file_name in os.listdir(dataset_source+'results/'):
    if file_name == 'base.txt':
        continue
    dataset = pd.read_csv(dataset_source+'results/'+file_name)
    dataset['seal_likelihood'] = dataset['seal_likelihood'].fillna(0)

    # Split into train/test
    X_train, X_test, E_train, E_test, Y_train,Y_test = train_test_split(dataset[['asim_likelihood','seal_likelihood']], dataset['edge_exists_original'], dataset['edge_exists_modified'],
                                             test_size=0.3, random_state=4269, stratify=dataset['edge_exists_modified'])

    # 'Train', Train a decision tree with both likelihood functions
    #clf = OneClassSVM()
    clf = tree.DecisionTreeClassifier()
    clf = clf.fit(X_train, Y_train)

    # Evaluate F1 score all over
    train_f1_score = f1_score(clf.predict(X_train).astype('float64'), Y_train)
    test_f1_score = f1_score(clf.predict(X_test).astype('float64'), Y_test)
    # Comparing to the original graph (just the test ones)
    exists_test_f1_score = f1_score(clf.predict(X_test).astype('float64'), E_test)
    # Comparing to the original graph (ALL OF THEM!)
    exists_all_f1_score = f1_score(clf.predict(pd.concat([X_train, X_test])).astype('float64'), 
                               np.concatenate((E_train,E_test)))

    # To save the results of all the training
    results_rows.append({'file_name':file_name, 'train_f1_score':train_f1_score, 'test_f1_score':test_f1_score, 'exists_f1_score':exists_test_f1_score, 'exists_all_f1_score':exists_all_f1_score})

    dataset['tree_result'] = clf.predict(dataset[['asim_likelihood','seal_likelihood']])
    
    if file_name != 'base.txt':
        # Creating graph only adding new edges, with weight 1 for all
        selected_edges = dataset[(dataset['edge_exists_modified']==1) | ((dataset['tree_result']>0.5))].copy()
        selected_edges[['from','to']].to_csv(dataset_source+'resulting_graphs/add_edges/'+file_name, sep=' ', header=None, index=False)

    else:
        # Creating graph as the base graph only
        selected_edges = dataset[(dataset['edge_exists_modified']==1)].copy()
        selected_edges[['from','to']].to_csv(dataset_source+'resulting_graphs/add_edges/'+file_name, sep=' ', header=None, index=False)

threshold_results = pd.DataFrame(results_rows)
threshold_results.to_csv(dataset_source+'thresholds_scores.csv', index=False)


threshold_results
